requirements.txt
# Core
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# GCP
google-cloud-documentai>=2.20.0
google-cloud-aiplatform>=1.38.0
google-genai>=0.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Notebook (SME Review Interface)
jupyter>=1.0.0
notebook>=7.0.0
ipywidgets>=8.0.0
